requirements.txt
aniso8601==8.0.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1596826740294/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1600298409949/work
backcall==0.2.0
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1601138718136/work
bleach @ file:///tmp/build/80754af9/bleach_1600439572647/work
blis==0.4.1
Bottleneck==1.3.2
brotlipy==0.7.0
catalogue==1.0.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1600681203656/work
chardet @ file:///D:/bld/chardet_1600358865323/work
click==7.1.2
colorama==0.4.3
cryptography @ file:///D:/bld/cryptography_1600960356847/work
cycler==0.10.0
cymem @ file:///D:/bld/cymem_1591801266669/work
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fastai==1.0.61
fastprogress @ file:///home/jhoward/anaconda3/conda-bld/fastprogress_1597927594295/work
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
future==0.18.2
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593431910144/work
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1596472773465/work
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1600976720582/work
jupyterthemes==0.20.0
kiwisolver==1.2.0
lesscpy==0.14.0
lshashpy3==0.0.8
MarkupSafe==1.1.1
matplotlib @ file:///C:/ci/matplotlib-base_1597858295275/work
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.2
nb-conda-kernels @ file:///C:/ci/nb_conda_kernels_1600445421227/work
nbclient @ file:///tmp/build/80754af9/nbclient_1601059699549/work
nbconvert @ file:///C:/ci/nbconvert_1601062655246/work
nbformat==5.0.7
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1600977170767/work
nmslib==2.0.6
notebook @ file:///C:/ci/notebook_1596840158498/work
numexpr @ file:///D:/bld/numexpr_1591280480683/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233948451/work
nvidia-ml-py3==7.352.0
olefile==0.46
packaging==20.4
pandas @ file:///D:/bld/pandas_1599601182795/work
pandocfilters==1.4.2
parso==0.7.0
pathlib==1.0.1
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1594299032796/work
plac==1.1.3
ply==3.11
preshed @ file:///D:/bld/preshed_1591802317560/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598885458782/work
psutil==5.7.2
pybind11==2.5.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1592425495151/work
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1597680516047/work
spacy @ file:///D:/bld/spacy_1594659357946/work
srsly @ file:///D:/bld/srsly_1589222535061/work
terminado==0.8.3
testpath==0.4.4
thinc @ file:///D:/bld/thinc_1590400627616/work
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1601259122622/work
traitlets==4.3.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1595434816409/work
wasabi @ file:///home/conda/feedstock_root/build_artifacts/wasabi_1600272362626/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.1.1
